requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
dill==0.2.7.1
feedparser==5.2.1
firebase-admin==2.5.0
Flask==0.12.2
Flask-Login==0.4.1
Flask-SocketIO==2.9.3
future==0.16.0
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-firestore==0.28.0
google-cloud-storage==1.6.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
googlemaps==2.5.1
grpcio==1.7.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
ply==3.8
pyasn1==0.3.7
pyasn1-modules==0.1.5
pyreadline==2.1
python-engineio==2.0.1
python-firebase==1.2
python-socketio==1.8.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
